+++ flink-connectors/flink-connector-kafka-0.10/src/main/resources/META-INF/services/connectors.TableConnectorFactory
