+++ flink-connectors/flink-connector-kafka-0.8/src/main/resources/META-INF/services/connectors.TableConnectorFactory
