+++ flink-libraries/flink-sql-client/src/main/java/org/apache/flink/table/client/gateway/local/ExecutionContext.java
@@ -49,8 +49,8 @@ import org.apache.flink.table.client.config.Source;
 import org.apache.flink.table.client.config.SourceSink;
 import org.apache.flink.table.client.gateway.SessionContext;
 import org.apache.flink.table.client.gateway.SqlExecutionException;
-import org.apache.flink.table.connector.TableSinkFactoryService;
-import org.apache.flink.table.connector.TableSourceFactoryService;
+import org.apache.flink.table.connectors.TableSinkFactoryService;
+import org.apache.flink.table.connectors.TableSourceFactoryService;
 import org.apache.flink.table.sinks.TableSink;
 import org.apache.flink.table.sources.TableSource;
 import org.apache.flink.util.FlinkException;
