+++ flink-libraries/flink-sql-client/src/test/java/org/apache/flink/table/client/gateway/utils/TestTableSinkFactory.java
@@ -23,7 +23,7 @@ import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.api.Types;
 import org.apache.flink.table.client.gateway.local.DependencyTest;
-import org.apache.flink.table.connector.TableConnectorFactory;
+import org.apache.flink.table.connectors.TableConnectorFactory;
 import org.apache.flink.table.descriptors.DescriptorProperties;
 import org.apache.flink.table.descriptors.TableDescriptorValidator;
 import org.apache.flink.table.sinks.AppendStreamTableSink;
@@ -48,7 +48,7 @@ import static org.apache.flink.table.descriptors.SchemaValidator.SCHEMA_TYPE;
 public class TestTableSinkFactory implements TableConnectorFactory<TableSink<Row>> {
 
 	@Override
-	public String tableType() {
+	public String getType() {
 		return TableDescriptorValidator.TABLE_TYPE_VALUE_SINK();
 	}
