+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/descriptors/StreamTableSourceDescriptor.scala
@@ -19,7 +19,7 @@
 package org.apache.flink.table.descriptors
 
 import org.apache.flink.table.api.{StreamTableEnvironment, Table, TableException, ValidationException}
-import org.apache.flink.table.connector.TableSourceFactoryService
+import org.apache.flink.table.connectors.TableSourceFactoryService
 import org.apache.flink.table.sources.{StreamTableSource, TableSource}
 
 /**
