+++ flink-libraries/flink-table/src/test/resources/META-INF/services/connectors.TableConnectorFactory
@@ -13,5 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-org.apache.flink.table.connector.TestTableSinkFactory
-org.apache.flink.table.connector.TestTableSourceFactory
+connectors.TestTableSinkFactory
+connectors.TestTableSourceFactory
