+++ flink-libraries/flink-table/src/test/scala/org/apache/flink/table/connectors/TestTableSourceFactory.scala
@@ -16,18 +16,22 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.connector
+package org.apache.flink.table.connectors
 
 import java.util
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.table.descriptors.ConnectorDescriptorValidator._
-import org.apache.flink.table.descriptors.FormatDescriptorValidator._
+import org.apache.flink.table.api.TableSchema
+import org.apache.flink.table.descriptors.ConnectorDescriptorValidator.{CONNECTOR_PROPERTY_VERSION, CONNECTOR_TYPE}
+import org.apache.flink.table.descriptors.FormatDescriptorValidator.{FORMAT_PROPERTY_VERSION, FORMAT_TYPE}
 import org.apache.flink.table.descriptors.TableDescriptorValidator
-import org.apache.flink.table.sinks.TableSink
+import org.apache.flink.table.sources.TableSource
 import org.apache.flink.types.Row
 
-class TestTableSinkFactory extends TableConnectorFactory[TableSink[Row]] {
+/**
+  * Table source factory for testing.
+  */
+class TestTableSourceFactory extends TableConnectorFactory[TableSource[Row]] {
 
   override def requiredContext(): util.Map[String, String] = {
     val context = new util.HashMap[String, String]()
@@ -48,24 +52,16 @@ class TestTableSinkFactory extends TableConnectorFactory[TableSink[Row]] {
     properties
   }
 
-  override def create(properties: util.Map[String, String]): TableSink[Row] = {
+  override def create(properties: util.Map[String, String]): TableSource[Row] = {
     if (properties.get("failing") == "true") {
       throw new IllegalArgumentException("Error in this factory.")
     }
-    new TableSink[Row] {
-      override def getOutputType: TypeInformation[Row] = throw new UnsupportedOperationException()
+    new TableSource[Row] {
+      override def getTableSchema: TableSchema = throw new UnsupportedOperationException()
 
-      override def getFieldNames: Array[String] = throw new UnsupportedOperationException()
-
-      override def getFieldTypes: Array[TypeInformation[_]] =
-        throw new UnsupportedOperationException()
-
-      override def configure(fieldNames: Array[String],
-                             fieldTypes: Array[TypeInformation[_]]): TableSink[Row] =
-        throw new UnsupportedOperationException()
+      override def getReturnType: TypeInformation[Row] = throw new UnsupportedOperationException()
     }
   }
 
-  override def tableType(): String = TableDescriptorValidator.TABLE_TYPE_VALUE_SINK
+  override def getType(): String = TableDescriptorValidator.TABLE_TYPE_VALUE_SOURCE
 }
-
